+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## **Unreleased**
 **Hearthstone**
+- Added King Of The Underbelly Band to the decklist.
 - Fixed cases where Commander Beatrix would not add cards to the deck.
 - Fixed Arenasmith Hero Pick Overlay positioning.